Removed duplicates for NumberInput
