+++ ui/src/components/ColumnTitle/FilterField.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { Checkbox, Icon, Input } from 'antd'
 import { CheckboxChangeEvent } from 'antd/lib/checkbox'
-import { NumberInput } from '@cxboxComponents'
+import NumberInput from '@components/ui/NumberInput/NumberInput'
 import CheckboxFilter from './CheckboxFilter/CheckboxFilter'
 import RangePicker from './RangePicker'
 import DatePicker from './DatePicker'
 import NumberRangeFilter from './components/NumberRangeFilter/NumberRangeFilter'
 import { getFormat } from '@utils/date'
-import { NumberTypes } from '@cxboxComponents/ui/NumberInput/formaters'
+import { NumberTypes } from '@components/ui/NumberInput/formaters'
 import { DataValue, FieldType, RowMetaField, WidgetListField, WidgetMeta } from '@cxbox-ui/core'
 import { DateFieldTypes } from '@interfaces/date'
 import { AppNumberFieldMeta, CustomFieldTypes } from '@interfaces/widget'
