+++ ui/src/components/TemplatedTitle/utils.tsx
@@ -5,7 +5,7 @@ import cn from 'classnames'
 import moment from 'moment'
 import { getFormat as getDateFormat } from '@utils/date'
 import { FieldType } from '@cxbox-ui/core'
-import { NumberInputFormat } from '@cxboxComponents/ui/NumberInput/formaters'
+import { NumberInputFormat } from '@components/ui/NumberInput/formaters'
 import { DateFieldMeta, DateTimeFieldMeta, DateTimeWithSecondsFieldMeta } from '@cxbox-ui/core'
 import { ITimePickerFieldMeta } from '../../fields/TimePicker/TimePickerField'
 import { AppNumberFieldMeta, CustomFieldTypes } from '@interfaces/widget'
