+++ ui/src/components/ui/NumberInput/NumberInput.less
