+++ ui/src/components/ui/NumberInput/NumberInput.tsx
@@ -1,9 +1,10 @@
 import React, { RefAttributes } from 'react'
-import { Input, Input as InputRef } from 'antd'
-import { fractionsRound, NumberInputFormat, NumberTypes } from '@cxboxComponents/ui/NumberInput/formaters'
+import { Input } from 'antd'
 import { InputProps } from 'antd/es/input'
 import { BaseFieldProps } from '@components/Field/Field'
+import { fractionsRound, NumberInputFormat, NumberTypes } from './formaters'
 import ReadOnlyField from '@components/ui/ReadOnlyField/ReadOnlyField'
+import styles from './NumberInput.less'
 
 export interface NumberInputProps extends BaseFieldProps {
     onChange?: (value: number) => void
@@ -23,8 +24,8 @@ export interface NumberInputProps extends BaseFieldProps {
  * @category Components
  */
 const NumberInput: React.FunctionComponent<NumberInputProps> = props => {
-    const { type, value, digits = 0, nullable, maxInput, forceFocus, onChange, onChangeInputValue, ...rest } = props
-    const inputRef = React.useRef<InputRef>(null)
+    const { type, value, digits, nullable, maxInput, onChange, onChangeInputValue } = props
+    const inputRef = React.useRef<Input>(null)
 
     const currency = type === NumberTypes.money && props.currency
 
@@ -54,7 +55,7 @@ const NumberInput: React.FunctionComponent<NumberInputProps> = props => {
             if (nullable && text === '') {
                 return null
             }
-            return fractionsRound(Number(normalizeValueFormat(text)), digits)
+            return fractionsRound(Number(normalizeValueFormat(text)), digits as number)
         },
         [nullable, digits]
     )
@@ -86,14 +87,14 @@ const NumberInput: React.FunctionComponent<NumberInputProps> = props => {
             if (inputRef.current?.input) {
                 const target = inputRef.current.input
                 const targetValue = target.value
-                const selectionStart = target.selectionStart
-                const selectionEnd = target.selectionEnd
+                const selectionStart = target.selectionStart as number
+                const selectionEnd = target.selectionEnd as number
 
                 const unformatedValue = unformatValue(targetValue)
                 setMode('edit')
 
                 target.value = unformatedValue
-                const selection = getUnformatedValueSelection(targetValue, selectionStart as number, selectionEnd as number)
+                const selection = getUnformatedValueSelection(targetValue, selectionStart, selectionEnd)
                 target.setSelectionRange(selection[0], selection[1])
             }
         }, 0)
@@ -127,11 +128,12 @@ const NumberInput: React.FunctionComponent<NumberInputProps> = props => {
         [props.maxInput, valueText]
     )
 
-    const extendedProps: InputProps & RefAttributes<InputRef> = {
-        ...rest,
+    const extendedProps: InputProps & RefAttributes<Input> = {
+        ...props,
         style: {
             backgroundColor: props.backgroundColor || '#fff'
         },
+        className: styles.container,
         addonAfter: currency,
         onChange: handleOnChange,
         onBlur: handleOnBlur,
@@ -140,7 +142,7 @@ const NumberInput: React.FunctionComponent<NumberInputProps> = props => {
         type: 'text',
         ref: inputRef,
         onKeyPress: onKeyPress,
-        autoFocus: forceFocus
+        autoFocus: props.forceFocus
     }
 
     if (props.readOnly) {
