+++ ui/src/components/ui/NumberInput/formaters.ts
