+++ ui/src/cxboxComponents/index.ts
@@ -1,7 +1,2 @@
-/**
- * @category Components
- */
-export { default as NumberInput } from '@cxboxComponents/ui/NumberInput/NumberInput'
-
 // Containers
 export { default as View } from './View/View'
